Added Version 3 of IPAN Militia scenario and updated documentation.
Version 3 of the IPAN militia scenario adds a new agent 'GordanianGovt' and adds to the scenario to create more paths for uncertain reasoning. The scenario is appended to as follows: The Gordanian Government has been approached by the leader of the Red militia faction promising to cooperate with the relief operations in an increased role in the political arena which has been agreed to by the Gordanian Government. The promise of cooperation with relief operations may indicate that the Red militia will suspend violence.